Fixed bug 999079 .This is caused by the preg_replace functions in the _exapandlinks function.
I tried to decode which regex was causing this but was
unable to, so I just added an additional regex to remove the
trailing slash from the URI before the page is concatenated
onto the end of it.
My new regex is :

		$match = preg_replace("|/$|","",$match);

This isn't elegant but it works. If anyone wants to
determine which of the 6 other regexes being run is causing
the double slashes, I'd be happy to fix it and take out my
additional line of code.